requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Talisman==1.1.0
Flask-Limiter==4.0.0
PyJWT==2.10.1
python-dotenv==1.0.0
Werkzeug==3.1.4

